Versione 0.0.750
Corretta a gestione del login/logout dell'account.
Iniziata la fase di transizione dalla vecchia gestione del DB alla nuova gestione con ORMLite.

+++ trunk/Interventix/src/com/federicocolantoni/projects/interventix/task/GetUtenteAsyncTack.java
@@ -45,7 +45,7 @@ public class GetUtenteAsyncTack extends AsyncTask<Long, Void, Utente> {
 	    utente.cognome = (cursor.getString(cursor.getColumnIndex(UtenteDB.Fields.COGNOME)));
 	    utente.conflitto = (cursor.getInt(cursor.getColumnIndex(UtenteDB.Fields.CONFLITTO)) == 1 ? true : false);
 	    utente.email = (cursor.getString(cursor.getColumnIndex(UtenteDB.Fields.EMAIL)));
-	    utente.idutente = (cursor.getInt(cursor.getColumnIndex(UtenteDB.Fields.ID_UTENTE)));
+	    utente.idutente = (cursor.getLong(cursor.getColumnIndex(UtenteDB.Fields.ID_UTENTE)));
 	    utente.nome = (cursor.getString(cursor.getColumnIndex(UtenteDB.Fields.NOME)));
 	    utente.revisione = (cursor.getLong(cursor.getColumnIndex(UtenteDB.Fields.REVISIONE)));
 	    utente.tipo = (cursor.getString(cursor.getColumnIndex(UtenteDB.Fields.TIPO)));
